# [R4R]txindex/kv: Fsync data to disk immediately after receiving it 