@ebuchman 2019-11-19
@melekes what is that benchmark on ? Individual `Index` calls or `AddBatch` calls? The former we don't seem to actually use other than in tests, the later I would expect much lesser performance hit from.

Regardless this is an important fix - thanks @guagualvcha . Seems also independent from https://github.com/tendermint/tendermint/issues/3211 which we'll need to address as well.