@melekes 2019-11-20
> @melekes what is that benchmark on ? Individual Index calls or AddBatch calls?

AddBatch calls

https://github.com/tendermint/tendermint/blob/97222eb90fc4b2882fe1859a665f2c84764c4b58/state/txindex/kv/kv_test.go#L314-L353