@ebuchman 2019-11-20
Hmm. Is the first number in `BenchmarkTxIndex2000-2` the number of txs? The cost is pretty high for the smaller nums but seems it's amortized for larger ones so maybe it's ok. Crazy that the cost is like 70x for the first one and otherwise only around ~1.2x or so for the others