@melekes 2019-11-20
> Is the first number in BenchmarkTxIndex2000-2 the number of txs?

Yes